requirements.txt
faiss-cpu==1.10.0
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.41
langchain-text-splitters==0.3.6
ollama==0.4.7
pypdf==5.3.1
sentence-transformers==3.4.1
